Merger of CABB development branch onto trunk.

+++ inc/miriad.h
@@ -256,8 +256,8 @@ void xysetpl_c (int tno, int naxis, Const int *axes);
 /* maskio.c */
 char *mkopen_c (int tno, char *name, char *status);
 void mkclose_c (char *handle);
-int  mkread_c  (char *handle, int mode, int *flags, int offset, int n, int nsize);
-void mkwrite_c (char *handle, int mode, Const int *flags, int offset, int n, int nsize);
+int  mkread_c  (char *handle, int mode, int *flags, off_t offset, int n, int nsize);
+void mkwrite_c (char *handle, int mode, Const int *flags, off_t offset, int n, int nsize);
 void mkflush_c (char *handle);
